Backend user validation

+++ frontend/src/queries/groupsQueries.ts
@@ -1,24 +1,27 @@
+import { useContext } from "react";
 import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
 import type GroupScheme from "../types/groupScheme";
 import { ShowError } from "../components/ShowError";
 import { cleanNotifications } from "@mantine/notifications";
 import { formatLastMessageDate, formatFullDate } from "../utils/FormatDate";
+import { CurrentUserContext } from "../contexts/CurrentUserContext";
 
 const API_URL = import.meta.env.VITE_API_URL;
 
 export function getGroups() {
+    const { currentUser } = useContext(CurrentUserContext);
+
     return useQuery({
         queryKey: ["groups"],
         queryFn: async () => {
-            const response = await fetch(`${API_URL}/groups`, {method: "GET"});
-            const result = await response.json();
+            if (!currentUser) return ShowError("You must be logged in to view groups.");
 
-            if (!response.ok) {
-                const error = new Error("Error getting groups: " + result.title);
-                console.log(error.message);
-                ShowError(error.message);
-                throw error;
-            }
+            const response = await fetch(`${API_URL}/groups`, {
+                method: "GET",
+                headers: { 'userId': currentUser.id!.toString() }
+            });
+            const result = await response.json();
+            if (!response.ok) return ShowError("Error getting groups: " + result.title);
 
             const groupsWithLocalTime = result.map((group: GroupScheme) => ({
                 ...group,
@@ -34,23 +37,21 @@ export function getGroups() {
 }
 
 export function createGroup() {
+    const { currentUser } = useContext(CurrentUserContext);
     const queryClient = useQueryClient();
 
     return useMutation({
         mutationFn: async (newGroup: GroupScheme) => {
+            if (!currentUser) return ShowError("You must be logged in to create groups.");
+
             const response = await fetch(`${API_URL}/groups`, {
                 method: 'POST',
-                headers: { 'Content-Type': 'application/json' },
+                headers: { 'Content-Type': 'application/json', 'userId': currentUser.id!.toString() },
                 body: JSON.stringify(newGroup),
             });
             const result = await response.json();
+            if (!response.ok) return ShowError("Error creating group: " + result.title);
 
-            if (!response.ok) {
-                const error = new Error("Error creating group: " + result.title);
-                console.log(error.message);
-                ShowError(error.message);
-                throw error;
-            }
             console.log("Group created:", result);
             queryClient.invalidateQueries({ queryKey: ['groups'] });
             cleanNotifications();
@@ -60,23 +61,21 @@ export function createGroup() {
 }
 
 export function editGroup() {
+    const { currentUser } = useContext(CurrentUserContext);
     const queryClient = useQueryClient();
 
     return useMutation({
         mutationFn: async (updatedGroup: GroupScheme) => {
+            if (!currentUser) return ShowError("You must be logged in to edit groups.");
+
             const response = await fetch(`${API_URL}/groups/${updatedGroup.id}`, {
                 method: 'PUT',
-                headers: { 'Content-Type': 'application/json' },
+                headers: { 'Content-Type': 'application/json', 'userId': currentUser.id!.toString() },
                 body: JSON.stringify(updatedGroup),
             });
             const result = await response.json();
+            if (!response.ok) return ShowError("Error updating group: " + result.title);
 
-            if (!response.ok) {
-                const error = new Error("Error updating group: " + result.title);
-                console.log(error.message);
-                ShowError(error.message);
-                throw error;
-            }
             console.log("Group updated:", result);
             queryClient.invalidateQueries({ queryKey: ['groups'] });
             cleanNotifications();
@@ -86,21 +85,20 @@ export function editGroup() {
 }
 
 export function deleteGroup() {
+    const { currentUser } = useContext(CurrentUserContext);
     const queryClient = useQueryClient();
 
     return useMutation({
         mutationFn: async (group: GroupScheme) => {
+            if (!currentUser) return ShowError("You must be logged in to delete groups.");
+
             const response = await fetch(`${API_URL}/groups/${group.id}`, {
                 method: 'DELETE',
+                headers: { 'userId': currentUser.id!.toString() },
             });
             const result = await response.json();
+            if (!response.ok) return ShowError("Error deleting group: " + result.title);
 
-            if (!response.ok) {
-                const error = new Error("Error deleting group: " + result.title);
-                console.log(error.message);
-                ShowError(error.message);
-                throw error;
-            }
             console.log("Group deleted:", result);
             queryClient.invalidateQueries({ queryKey: ['groups'] });
             cleanNotifications();
